main changes to pages components include:
-installed a plugin to orgrnize  html tags so code will be more
orgrianized/pretty

-regrading comments section:
relvant components were udpated. mainly
include improve of the design,  removal of redundent code and addition of polling method.
also contins the sending of post request to server with the comments and
reply data.
contains basic filtering for the comments data (not done yet)

-added componenets for the user's last viewed  docuemnts page. currently
contains 5 documents in a row as if optimaly fits in the page.

-added an Intro components that will show in the home page when the user
first logged in (will be updated with further data from server later)

-added route in navbar for user profile page
the profile view contains a components showing the logged in user
personal details. thid contails also a basic pop modal for
uploding/changing the user's image (not done yet)
still missing function for changing regular profile  details (username and more)